requirements.txt
python-telegram-bot==20.7
pyrebase4==4.7.1
python-dotenv==1.0.0
pydantic==2.5.0
telegram